+++ test/Middleware/WhoopsErrorResponseGeneratorTest.php
@@ -7,6 +7,7 @@
 
 namespace ZendTest\Expressive\Middleware;
 
+use Fig\Http\Message\StatusCodeInterface as StatusCode;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Http\Message\ResponseInterface;
@@ -58,7 +59,9 @@ class WhoopsErrorResponseGeneratorTest extends TestCase
         $this->request->getAttribute('originalUri', false)->shouldNotBeCalled();
         $this->request->getAttribute('originalRequest', false)->shouldNotBeCalled();
 
+        $this->response->withStatus(StatusCode::STATUS_INTERNAL_SERVER_ERROR)->will([$this->response, 'reveal']);
         $this->response->getBody()->will([$this->stream, 'reveal']);
+        $this->response->getStatusCode()->willReturn(StatusCode::STATUS_INTERNAL_SERVER_ERROR);
 
         $this->stream->write('WHOOPS')->shouldBeCalled();
 
@@ -72,7 +75,7 @@ class WhoopsErrorResponseGeneratorTest extends TestCase
 
     public function testAddsRequestMetadataToWhoopsPrettyPageHandler()
     {
-        $error = new RuntimeException();
+        $error = new RuntimeException('STATUS_INTERNAL_SERVER_ERROR', StatusCode::STATUS_INTERNAL_SERVER_ERROR);
 
         $handler = $this->prophesize(PrettyPageHandler::class);
         $handler
@@ -101,6 +104,8 @@ class WhoopsErrorResponseGeneratorTest extends TestCase
         $this->request->getQueryParams()->willReturn([]);
         $this->request->getParsedBody()->willReturn([]);
 
+        $this->response->withStatus(StatusCode::STATUS_INTERNAL_SERVER_ERROR)->will([$this->response, 'reveal']);
+        $this->response->getStatusCode()->willReturn(StatusCode::STATUS_INTERNAL_SERVER_ERROR);
         $this->response->getBody()->will([$this->stream, 'reveal']);
 
         $this->stream->write('WHOOPS')->shouldBeCalled();
@@ -115,7 +120,7 @@ class WhoopsErrorResponseGeneratorTest extends TestCase
 
     public function testJsonContentTypeResponseWithJsonResponseHandler()
     {
-        $error = new RuntimeException();
+        $error = new RuntimeException('STATUS_NOT_IMPLEMENTED', StatusCode::STATUS_NOT_IMPLEMENTED);
 
         $handler = $this->prophesize(JsonResponseHandler::class);
 
@@ -137,6 +142,8 @@ class WhoopsErrorResponseGeneratorTest extends TestCase
         $this->request->getParsedBody()->willReturn([]);
 
         $this->response->withHeader('Content-Type', 'application/json')->will([$this->response, 'reveal']);
+        $this->response->withStatus(StatusCode::STATUS_NOT_IMPLEMENTED)->will([$this->response, 'reveal']);
+        $this->response->getStatusCode()->willReturn(StatusCode::STATUS_NOT_IMPLEMENTED);
         $this->response->getBody()->will([$this->stream, 'reveal']);
 
         $this->stream->write('error')->shouldBeCalled();